Writes videos to a temporary file first, renames the file at the end
This is related to groupon/Selenium-Grid-Extras #292.

Effectively, the code which downloads the file has no notion of whether the file is an in-progress video or not.  Because of this, "eager consumers" can attempt to construct the download URL themselves (since they are in posession of the session ID), which will return a 200 but download an incomplete file.

Instead, we change the code so that it writes the video to {sessionId}.temp, and at the end, once the video is fully written, we rename the file to it's final destination of {sessionId}.mp4.  This means that the downloader will only see the mp4 file in cases where it is fully written.