Do some memory cleanup.
Humble-video claims that it should detect and clean resources. Not sure about the Graphics cleanup how important it is.
I ran jcmd pid GC.run and the memory was reduced on a process with the fixes and on a process without the fixes, so it shouldn't be a problem after all.